services/transcriber/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
# Pin Whisper version to ensure stable DecodingOptions signature.
# Update deliberately after testing (e.g. pip install --upgrade openai-whisper && record new version).
openai-whisper==20250625
# PyTorch and related libraries
# Ensure to use the PyTorch index URL for installation:
# pip install --upgrade torch torchaudio --index-url https://download.pytorch.org/whl/cpu
#torch==2.4.0
torch
#torchaudio==2.4.0
torchaudio
numpy
aio_pika==9.4.3
python-multipart==0.0.9
python-dotenv==1.0.1
pytz==2023.3  # Add pytz for timezone support
